fix several issues
